+++ front/react-app/src/Molecules/AlarmPresetsListItem.tsx
@@ -1,13 +1,15 @@
+import { useContext } from "react"
+import { AlarmPresetsContext } from "../providers/AlarmPresetsProvider"
 import { AlarmPresetsListItemType } from "../types/types"
 import { taskConverter } from "./Form"
 
 type AlarmPresetsListItemProps = {
-  alarmPresets: AlarmPresetsListItemType[],
+  correctPreset: AlarmPresetsListItemType,
   setCorrectPreset: React.Dispatch<React.SetStateAction<AlarmPresetsListItemType>>
 }
 
-export const AlarmPresetsListItem = ({ alarmPresets, setCorrectPreset } : AlarmPresetsListItemProps): JSX.Element => {
-
+export const AlarmPresetsListItem = ({ correctPreset, setCorrectPreset } : AlarmPresetsListItemProps): JSX.Element => {
+  const { alarmPresets } = useContext(AlarmPresetsContext);
   const onClickListItem = (id: number) => {
     setCorrectPreset({
       id: alarmPresets[id].id,
@@ -20,8 +22,13 @@ export const AlarmPresetsListItem = ({ alarmPresets, setCorrectPreset } : AlarmP
   return (
     <>
       { alarmPresets.map((alarmPreset: AlarmPresetsListItemType, index: number) => {
+          let selectedBackGroundCSS: string = '';
+          if (correctPreset.id === alarmPreset.id) {
+            selectedBackGroundCSS = 'bg-red-200';
+          }
+
           return (
-            <li key={ index } className="w-full flex justify-between hover:bg-gray-300 cursor-pointer" onClick={ () => onClickListItem( index ) }>
+            <li key={ index } className={ `w-full flex justify-between hover:bg-red-100 cursor-pointer ${ selectedBackGroundCSS }` } onClick={ () => onClickListItem( index ) }>
               <p className="w-1/3 border truncate px-2">{ alarmPreset.presetName }</p>
               <p className="w-1/3 border truncate px-2">{ alarmPreset.wakeAt?.format('HH:mm') }</p>
               <p className="w-1/3 border truncate px-2">{ taskConverter(alarmPreset.task) }</p>
